+++ src/assets/ba_data/python/bascenev1/_lobby.py
@@ -339,10 +339,10 @@ class Chooser:
         if inputdevice.is_controller_app and '_random' in profilenames:
             return profilenames.index('_random')
 
-        # If its a client connection, for now just force
-        # the account profile if possible.. (need to provide a
-        # way for clients to specify/remember their default
-        # profile on remote servers that do not already know them).
+        # If its a client connection, for now just force the account
+        # profile if possible. (need to provide a way for clients to
+        # specify/remember their default profile on remote servers that
+        # do not already know them).
         if inputdevice.is_remote_client and '__account__' in profilenames:
             return profilenames.index('__account__')