+++ tools/batools/build.py
@@ -527,7 +527,7 @@ def _get_server_config_template_toml(projroot: str) -> str:
     cfg.clean_exit_minutes = 60
     cfg.unclean_exit_minutes = 90
     cfg.idle_exit_minutes = 20
-    cfg.admins = ['pb-yOuRAccOuNtIdHErE', 'pb-aNdMayBeAnotherHeRE']
+    cfg.admins = ['a-YOUR-ID-HERE', 'a-ANOTHER-ID-HERE']
     cfg.protocol_version = 35
     cfg.session_max_players_override = 8
     cfg.playlist_inline = []