+++ tools/efrotools/efrocache.py
@@ -51,8 +51,8 @@ class CacheMetadata:
     executable: Annotated[bool, IOAttrs('e')]
 
 
-g_cache_prefix_noexec: bytes | None = None
-g_cache_prefix_exec: bytes | None = None
+_g_cache_prefix_noexec: bytes | None = None
+_g_cache_prefix_exec: bytes | None = None
 
 
 def get_local_cache_dir() -> str:
@@ -664,8 +664,8 @@ def _cache_prefix_for_file(fname: str) -> bytes:
     # pylint: disable=global-statement
     from efrotools.util import is_wsl_windows_build_path
 
-    global g_cache_prefix_exec
-    global g_cache_prefix_noexec
+    global _g_cache_prefix_exec
+    global _g_cache_prefix_noexec
 
     # We'll be calling this a lot when checking existing files, so we
     # want it to be efficient. Let's cache the two options there are at
@@ -689,21 +689,21 @@ def _cache_prefix_for_file(fname: str) -> bytes:
         executable = False
 
     if executable:
-        if g_cache_prefix_exec is None:
+        if _g_cache_prefix_exec is None:
             metadata = dataclass_to_json(
                 CacheMetadata(executable=True)
             ).encode()
             assert len(metadata) < 256
-            g_cache_prefix_exec = (
+            _g_cache_prefix_exec = (
                 CACHE_HEADER + len(metadata).to_bytes() + metadata
             )
-        return g_cache_prefix_exec
+        return _g_cache_prefix_exec
 
     # Ok; non-executable it is.
     metadata = dataclass_to_json(CacheMetadata(executable=False)).encode()
     assert len(metadata) < 256
-    g_cache_prefix_noexec = CACHE_HEADER + len(metadata).to_bytes() + metadata
-    return g_cache_prefix_noexec
+    _g_cache_prefix_noexec = CACHE_HEADER + len(metadata).to_bytes() + metadata
+    return _g_cache_prefix_noexec
 
 
 def _check_warm_start_entry(entry: tuple[str, str]) -> None: