+++ CHANGELOG.md
@@ -1,4 +1,18 @@
-### 1.7.61 (build 22712, api 9, 2026-02-12)
+### 1.7.61 (build 22714, api 9, 2026-02-21)
+- Added scenev1 protocol 36, which enables V2 auth for servers. This allows
+  servers to receive authenticated V2 account info for all players before they
+  are allowed to join and fixes the spoofing vulnerabilities that V1 auth had.
+  V2 account ids look like 'a-XXX' whereas old V1 looked like 'pb-XXXX'. The
+  default protocol is still 33, but if you are running a server it is highly
+  recommended to set your protocol to 36 in your server config to enable this.
+- Clients will now wait for responses for up to 10 seconds when connecting to a
+  server and 30 seconds if contact is lost once connected. Hopefully this
+  reduces disconnects due to momentary network issues. Holler if this feels like
+  too long. Old values were 5 and 10 seconds respectively.
+- Improved efficiency of various low level logging calls in the C++ layer
+  (thanks std::string_view!).
+- Fixed an issue where clicks could sometimes be lost in the nearby-parties
+  browser.
 - Fixed party window sub-menus staying after closing the root window.
   (Thanks temp!)